feat(chat): update backend structure
