+++ examples/ui/.env.example
@@ -2,3 +2,4 @@ WORKSPACE_PATH=./workspace
 FASTAPI_RELOAD=false
 PANDA_AGI_KEY=
 TAVILY_API_KEY=
+E2B_API_KEY=
