+++ examples/ui/backend/routes/__init__.py
@@ -1,0 +1,3 @@
+"""
+Routes for the PandaAGI SDK API.
+"""
