+++ examples/ui/backend/routes/files.py
@@ -1,313 +1,40 @@
-import asyncio
-import json
+"""
+File routes for the PandaAGI SDK API.
+"""
 import logging
-import os
-
-# Import the agent and environment from the parent directory
-import sys
-import uuid
-from pathlib import Path
-from typing import AsyncGenerator, Dict, Optional, Union
-
-from dotenv import load_dotenv
-from fastapi import FastAPI, File, Form, HTTPException, Query, UploadFile
-from fastapi.middleware.cors import CORSMiddleware
-from fastapi.responses import FileResponse, StreamingResponse, Response
-from pydantic import BaseModel
-import aiohttp
-from fastapi import Depends, HTTPException, status
-from fastapi.security import HTTPBearer, HTTPAuthorizationCredentials
-
-
-# Load environment variables from .env file
-load_dotenv()
-
-sys.path.append(os.path.join(os.path.dirname(__file__), "..", ".."))
-
-from panda_agi import Agent, skill
-from panda_agi.client.models import (
-    BaseStreamEvent,
-    EventType,
-)
-from panda_agi.envs import E2BEnv
 import mimetypes
+import os
+import tempfile
+from pathlib import Path
+from typing import Optional
 
-# Configure logging with more explicit settings
-logging.basicConfig(
-    level=logging.DEBUG,
-    format="%(asctime)s - %(name)s - %(levelname)s - %(message)s",
-    handlers=[
-        logging.StreamHandler(sys.stdout),  # Ensure output goes to stdout
-        logging.FileHandler("api.log"),  # Also log to file
-    ],
-)
+from fastapi import APIRouter, File, Form, HTTPException, Query, UploadFile
+from fastapi.responses import FileResponse, Response
+
+from services.agent import get_or_create_agent
 
 logger = logging.getLogger("panda_agi_api")
-logger.setLevel(logging.DEBUG)
 
 # Get workspace path from environment variable with fallback
-# WORKSPACE_PATH = os.getenv(
-#     "WORKSPACE_PATH", os.path.join(os.path.dirname(__file__), "workspace")
-# )
-
 WORKSPACE_PATH = "/workspace"
 
-app = FastAPI(title="PandaAGI SDK API", version="1.0.0")
-
-# Auth bearer security scheme for validating tokens
-security = HTTPBearer()
-
-# Store active conversations - in production, this would be in a database
-active_conversations: Dict[str, Agent] = {}
+router = APIRouter(tags=["files"])
 
 
-# Add CORS middleware
-app.add_middleware(
-    CORSMiddleware,
-    allow_origins=["*"],
-    allow_credentials=True,
-    allow_methods=["*"],
-    allow_headers=["*"],
-)
-
-
-class AgentQuery(BaseModel):
-    query: str
-    conversation_id: Optional[str] = None
-    timeout: Optional[int] = None
-
-
-def should_render_event(event: Union[BaseStreamEvent, str]) -> bool:
-    """Check if event should be rendered - same logic as CLI"""
-    # Handle both BaseStreamEvent objects and string event types
-    if isinstance(event, BaseStreamEvent):
-        event_type = (
-            event.type.value if hasattr(event.type, "value") else str(event.type)
-        )
-    else:
-        event_type = event
-
-    # Skip redundant events
-    if event_type == EventType.WEB_NAVIGATION.value:
-        # Skip WEB_NAVIGATION since WEB_NAVIGATION_RESULT provides the same info + content
-        return False
-
-    # Skip task completion events - not needed in UI
-    if event_type == EventType.COMPLETED_TASK.value:
-        return False
-
-    # Skip agent connection success - not needed in UI
-    if event_type == EventType.AGENT_CONNECTION_SUCCESS.value:
-        return False
-
-    return True
-
-
-def truncate_long_content(data, max_length: int = 5000):
-    """
-    Recursively truncate long string values in a dictionary or list.
-
-    Args:
-        data: The data structure to truncate (dict, list, or any other type)
-        max_length: Maximum length for string values before truncation
-
-    Returns:
-        The data structure with truncated string values
-    """
-    if isinstance(data, dict):
-        return {
-            key: truncate_long_content(value, max_length) for key, value in data.items()
-        }
-    elif isinstance(data, list):
-        return [truncate_long_content(item, max_length) for item in data]
-    elif isinstance(data, str) and len(data) > max_length:
-        return data[:max_length] + "\n\n... [Content truncated]"
-    else:
-        return data
-
-
-def process_event_for_frontend(event) -> Optional[Dict]:
-    """Process an event for frontend consumption with type safety"""
-    try:
-        if isinstance(event, BaseStreamEvent):
-            event_type = (
-                event.type.value if hasattr(event.type, "value") else str(event.type)
-            )
-
-            # Get the event data
-            event_data = event.to_dict()
-
-            # Apply truncation to all string fields
-            event_data = truncate_long_content(event_data)
-
-            return {
-                "data": {
-                    "type": event_type,
-                    "payload": event_data,
-                    "timestamp": event.timestamp,
-                    "id": getattr(event, "id", None),
-                },
-            }
-        else:
-            # Handle legacy events with truncation
-            event_type = (
-                event.type.value if hasattr(event.type, "value") else str(event.type)
-            )
-
-            event_data = event.data
-            # Apply truncation to all string fields
-            event_data = truncate_long_content(event_data)
-
-            return {
-                "data": {
-                    "type": event_type,
-                    "payload": event_data,
-                    "timestamp": getattr(event, "timestamp", ""),
-                    "id": getattr(event, "id", None),
-                },
-            }
-    except Exception as e:
-        print(f"Error processing event: {e}")
-        return None
-
-
-def get_or_create_agent(conversation_id: Optional[str] = None) -> tuple[Agent, str]:
-    """Get existing agent or create new one for conversation"""
-    if conversation_id and conversation_id in active_conversations:
-        return active_conversations[conversation_id], conversation_id
-    
-    local_env: E2BEnv = E2BEnv(WORKSPACE_PATH)
-
-    @skill
-    async def deploy_python_server(port) -> str:
-        """
-        Deploys a simple Python HTTP server on the specified port and return new url
-
-        This function:
-        - Kills any process currently running on the given port.
-        - Starts a new Python HTTP server on that port.
-
-        Args:
-            port (int): The port number on which to start the server.
-
-        Returns:
-            str: URL of the running server (e.g., http://localhost:8000).
-        """
-        kill_cmd = f"fuser -k {port}/tcp || true"
-        start_cmd = f"nohup python -m http.server {port} > /dev/null 2>&1 &"
-        logger.debug(f"Executing deploy skill to start server at port: {port}")
-        await local_env.exec_shell(kill_cmd)
-        await local_env.exec_shell(start_cmd)
-        return await local_env.get_hosted_url(port)
-
-    agent = Agent(environment=local_env, skills=[deploy_python_server])
-    new_conversation_id = conversation_id or str(uuid.uuid4())
-    active_conversations[new_conversation_id] = agent
-
-    return agent, new_conversation_id
-
-
-async def event_stream(
-    query: str, conversation_id: Optional[str] = None
-) -> AsyncGenerator[str, None]:
-    """Stream agent events as Server-Sent Events"""
-    agent = None
-    actual_conversation_id = None
-
-    try:
-        # Get or create agent for this conversation
-        agent, actual_conversation_id = get_or_create_agent(conversation_id)
-
-        # Send conversation ID as first event
-        conversation_event = {
-            "data": {
-                "type": "conversation_started",
-                "payload": {"conversation_id": actual_conversation_id},
-                "timestamp": "",
-                "id": None,
-            }
-        }
-        yield f"data: {json.dumps(conversation_event)}\n\n"
-        await asyncio.sleep(0.01)
-
-        # Stream events
-        async for event in agent.run_stream(query):
-            # Apply filtering first
-            if not should_render_event(event):
-                continue
-
-            # Process event with type safety while maintaining frontend structure
-            event_dict = process_event_for_frontend(event)
-
-            if event_dict is None:
-                # Skip events that couldn't be processed
-                continue
-
-            # Format as SSE
-            yield f"data: {json.dumps(event_dict)}\n\n"
-
-    except Exception as e:
-        # Send error event
-        error_data = {
-            "data": {
-                "type": "error",
-                "payload": {"error": str(e)},
-                "timestamp": "",
-                "id": None,
-            },
-        }
-        yield f"data: {json.dumps(error_data)}\n\n"
-
-        # Clean up failed conversation
-        if actual_conversation_id and actual_conversation_id in active_conversations:
-            try:
-                await active_conversations[actual_conversation_id].disconnect()
-                del active_conversations[actual_conversation_id]
-            except Exception as cleanup_error:
-                print(f"Error cleaning up failed conversation: {cleanup_error}")
-
-
-@app.post("/agent/run")
-async def run_agent(query_data: AgentQuery):
-    """Run an agent with the given query and stream events"""
-    try:
-        logger.debug(f"Running agent with query: {query_data.query}")
-        logger.debug(f"Conversation ID: {query_data.conversation_id}")
-        return StreamingResponse(
-            event_stream(query_data.query, query_data.conversation_id),
-            media_type="text/event-stream",
-            headers={
-                "Cache-Control": "no-cache",
-                "Connection": "keep-alive",
-                "Access-Control-Allow-Origin": "*",
-            },
-        )
-    except Exception as e:
-        print(e)
-        raise HTTPException(status_code=500, detail=str(e))
-
-
-@app.delete("/conversation/{conversation_id}")
-async def end_conversation(conversation_id: str):
-    """End a conversation and clean up resources"""
-    if conversation_id in active_conversations:
-        try:
-            await active_conversations[conversation_id].disconnect()
-            del active_conversations[conversation_id]
-            return {"status": "conversation ended"}
-        except Exception as e:
-            raise HTTPException(
-                status_code=500, detail=f"Error ending conversation: {str(e)}"
-            )
-    else:
-        raise HTTPException(status_code=404, detail="Conversation not found")
-
-
-@app.post("/files/upload")
+@router.post("/files/upload")
 async def upload_files(
     file: UploadFile = File(...), conversation_id: Optional[str] = Form(None)
 ):
-    """Upload a file to the workspace using E2BEnv"""
+    """
+    Upload a file to the workspace using E2BEnv.
+    
+    Args:
+        file: The file to upload
+        conversation_id: Optional ID of the conversation
+        
+    Returns:
+        dict: Upload status and file information
+    """
     try:
         local_env = get_or_create_agent(conversation_id)[0].environment
         # Use the environment's workspace path
@@ -364,12 +91,21 @@ async def upload_files(
         raise HTTPException(status_code=500, detail=f"Upload failed: {str(e)}")
 
 
-@app.get("/{conversation_id}/files/download")
+@router.get("/{conversation_id}/files/download")
 async def download_file(
     conversation_id: str,
     file_path: str = Query(..., description="Path to the file to download"),
 ):
-    """Download a file from the workspace"""
+    """
+    Download a file from the workspace.
+    
+    Args:
+        conversation_id: ID of the conversation
+        file_path: Path to the file to download
+        
+    Returns:
+        FileResponse: The file to download
+    """
     try:
         local_env = get_or_create_agent(conversation_id)[0].environment
         logger.debug(f"Download request for file_path: '{file_path}'")
@@ -486,8 +222,6 @@ async def download_file(
                 logger.debug("Successfully converted HTML to PDF")
 
                 # Create a temporary PDF file
-                import tempfile
-
                 with tempfile.NamedTemporaryFile(
                     delete=False, suffix=".pdf"
                 ) as temp_pdf:
@@ -527,7 +261,6 @@ async def download_file(
             raise Exception(f"Failed to read file: {file_result.get('message', 'Unknown error')}")
         
         # Create a temporary file to serve
-        import tempfile
         with tempfile.NamedTemporaryFile(delete=False) as temp_file:
             # Ensure content is bytes
             content = file_result["content"]
@@ -549,10 +282,17 @@ async def download_file(
         raise HTTPException(status_code=500, detail=f"Error downloading file: {str(e)}")
 
 
-@app.get("/{conversation_id}/files/{file_path:path}")
+@router.get("/{conversation_id}/files/{file_path:path}")
 async def read_file(conversation_id: str, file_path: str):
     """
     Read a file from the E2B sandbox workspace and serve it directly.
+    
+    Args:
+        conversation_id: ID of the conversation
+        file_path: Path to the file to read
+        
+    Returns:
+        Response: The file content
     """
     try:
         local_env = get_or_create_agent(conversation_id)[0].environment
@@ -573,6 +313,7 @@ async def read_file(conversation_id: str, file_path: str):
 
         # Read file as binary to preserve any type
         read_res = await local_env.read_file(file_path, mode="rb")
+        print("Read file result: ", read_res)
         if read_res.get("status") != "success":
             detail = read_res.get("message", "Unknown error")
             raise HTTPException(status_code=500, detail=f"Error reading file: {detail}")
@@ -601,11 +342,19 @@ async def read_file(conversation_id: str, file_path: str):
         raise HTTPException(status_code=500, detail=f"Error reading file: {str(e)}")
 
 
-@app.get("/files/test-download")
+@router.get("/files/test-download")
 async def test_download_file(
     file_path: str = Query(..., description="Path to the file to download"),
 ):
-    """Test download endpoint"""
+    """
+    Test download endpoint.
+    
+    Args:
+        file_path: Path to the file to download
+        
+    Returns:
+        FileResponse or dict: The file to download or error message
+    """
     try:
         print(f"TEST DEBUG: Download request for file_path: '{file_path}'")
         print(f"TEST DEBUG: WORKSPACE_PATH: {WORKSPACE_PATH}")
@@ -633,90 +382,3 @@ async def test_download_file(
     except Exception as e:
         print(f"TEST DEBUG: Error: {e}")
         return {"error": str(e)}
-
-
-@app.get("/health")
-async def health_check():
-    """Health check endpoint"""
-    return {"status": "healthy"}
-
-
-@app.get("/")
-async def root():
-    """Root endpoint with API information"""
-    return {
-        "message": "PandaAGI SDK API",
-        "version": "1.0.0",
-        "endpoints": {
-            "POST /agent/run": "Run an agent with streaming events",
-            "DELETE /conversation/{conversation_id}": "End a conversation",
-            "POST /files/upload": "Upload a file to the workspace",
-            "GET /files/{file_path:path}": "Read a file from the workspace",
-            "GET /files/download": "Download a file from the workspace",
-            "GET /files/test-download": "Test download endpoint",
-            "GET /health": "Health check",
-            "GET /": "This endpoint",
-        },
-    }
-
-
-@app.get("/auth/github")
-async def github_auth(redirect_uri: Optional[str] = Query(None)):
-    """GitHub auth endpoint with optional redirect_uri"""
-    async with aiohttp.ClientSession() as session:
-        payload = {}
-        # if redirect_uri:
-        payload["redirect_uri"] = redirect_uri or "http://localhost:3001/authenticate"
-            
-        async with session.post(
-            "http://localhost:8000/public/auth/github",
-            json=payload
-        ) as resp:
-            response = await resp.json()
-            return response
-
-@app.get("/auth/validate")
-async def validate_auth(credentials: HTTPAuthorizationCredentials = Depends(security)):
-    """Validate authentication token by forwarding to backend service"""
-    token = credentials.credentials  # Extract the token from the bearer header
-    
-    async with aiohttp.ClientSession() as session:
-        # Pass the token in the Authorization header to the backend service
-        headers = {"X-Authorization": f"Bearer {token}"}
-        
-        async with session.get("http://localhost:8000/auth/validate", headers=headers) as resp:
-            if resp.status != 200:
-                raise HTTPException(
-                    status_code=resp.status,
-                    detail="Token validation failed"
-                )
-            
-            response = await resp.json()
-            return response
-
-
-@app.get("/debug-workspace")
-async def debug_workspace():
-    """Debug workspace path"""
-    import os
-
-    return {
-        "workspace_path": WORKSPACE_PATH,
-        "workspace_resolved": str(Path(WORKSPACE_PATH).resolve()),
-        "current_working_directory": os.getcwd(),
-        "file_exists": Path(WORKSPACE_PATH).exists(),
-    }
-
-
-if __name__ == "__main__":
-    import uvicorn
-
-    # Allow reload to be controlled by environment variable
-    # Defaults to True for development, can be set to False in production
-    reload_enabled = os.getenv("FASTAPI_RELOAD", "false").lower() in (
-        "true",
-        "1",
-        "yes",
-    )
-
-    uvicorn.run("e2b_main:app", host="0.0.0.0", port=8001, reload=reload_enabled)
