+++ examples/ui/backend/services/chat_env.py
@@ -1,0 +1,21 @@
+from panda_agi.envs import E2BEnv
+from panda_agi.envs.local_env import LocalEnv
+import os
+from typing import Dict, Any
+
+WORKSPACE_PATH = os.getenv(
+    "WORKSPACE_PATH", os.path.join(os.path.dirname(__file__), "workspace")
+)
+
+
+def get_env(metadata: Dict[str, Any] = None):
+    env = os.getenv("ENV", "local")
+    if env == "e2b":
+        # fetching existing for the conversation....
+        return E2BEnv("/workspace", metadata=metadata, timeout=1800)
+
+
+    print("Local environment", WORKSPACE_PATH)
+    return LocalEnv(WORKSPACE_PATH, metadata)
+
+
